Install the package and test requirements in one step in the travis setup
This should not have been removed. But it has effectively pointed out that the dance is not required in this scenario. The point of this line was to uninstall the development version (i.e. -e option to pip). This smoke tests the package manifest is accurately includes the files that need to be available in the packaged version.