Ensure license language matches metadata language
* Add new language constraints to the license parsing
* Clean up license parsing to make constraints easier to manage
* Update tests to check language constraints

There was a concern that license and metadata languages might become
out of sync. That is the case already: the en license is used in many
instances where md:language is not en. These changes allow that happen
for the sake of compatibility; however, this can be changed in the
future.

For now, only check to make sure that the license language and the
metadata language match when the license is localized (i.e. licenses
ending in /deed.xx).