1860 default license text AND url (#56)
* Return default license if license url is MIA;
return license text if url is localized

The default license is a temporary addition to help build user books.
Once we are done building user books, we can remove this fallback.

* Add parsing tests for default license and localization

* Add url to license type-or-version error for clarity

It's easier to find the error when you know what it looks like

* Ensure license language matches metadata language

1. Add new language constraints to the license parsing
2. Clean up license parsing to make constraints easier to manage
3. Update tests to check language constraints

There was a concern that license and metadata languages might become
out of sync. That is the case already: the en license is used in many
instances where md:language is not en. These changes allow that happen
for the sake of compatibility; however, this can be changed in the
future.

For now, only check to make sure that the license language and the
metadata language match when the license is localized (i.e. licenses
ending in /deed.xx).